requirements.txt
discord.py
pydle
aiosqlite